Removed Bootstrap
The mobile navigation bar was broken with Bootstrap, and we were really only using for the carousel. Bootstrap has been ditched in favor of the SCSS included in the Minima theme, and the Slick carousel library. Styling is now managed via SCSS rather than CSS.

Additionally, the volunteer form has been removed, as it is not clear if it ever worked. It doesn't seem to be working, and is probably not worth the effort of fixing at this time.